backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.0
python-multipart==0.0.9
python-dotenv==1.0.0
transformers==4.38.2
torch==2.2.1
sentence-transformers==2.7.0
pydantic==2.6.4
numpy==1.26.4
scikit-learn==1.4.1.post1
